+++ extension/observer/k8s_observer/config.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package k8s
+package k8s_observer
 
 import (
 	"github.com/open-telemetry/opentelemetry-collector/config/configmodels"
