+++ extension/observer/k8s_observer/go.mod
