+++ extension/observer/k8s_observer/handler.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package k8s
+package k8s_observer
 
 import (
 	"fmt"
