+++ extension/observer/k8s_observer/k8s_fixtures_test.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package k8s
+package k8s_observer
 
 import (
 	v1 "k8s.io/api/core/v1"
