+++ go.mod
@@ -63,7 +63,7 @@ replace github.com/open-telemetry/opentelemetry-collector-contrib/exporter/stack
 
 replace github.com/open-telemetry/opentelemetry-collector-contrib/extension/observer => ./extension/observer
 
-replace github.com/open-telemetry/opentelemetry-collector-contrib/extension/observer/k8s => ./extension/observer/k8s
+replace github.com/open-telemetry/opentelemetry-collector-contrib/extension/observer/k8s => ./extension/observer/k8s_observer
 
 replace github.com/open-telemetry/opentelemetry-collector-contrib/receiver/carbonreceiver => ./receiver/carbonreceiver
